added interface to allow objects to serialize themselves added interface for objects that know about their serializer options send publish/record start events before the first packet is received correctly decode published stream metadata -> Now Red5 supports being connected to from FME!
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@1760 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/RTMPHandler.java
@@ -276,7 +276,7 @@ public class RTMPHandler extends BaseRTMPHandler {
 											pc.setResult(getStatus(NC_CONNECT_SUCCESS));
 										}
 										// Measure initial roundtrip time after connecting
-										conn.getChannel((byte) 2).write(new Ping((short) 0, 0, -1));
+										conn.getChannel(2).write(new Ping((short) 0, 0, -1));
 										conn.startRoundTripMeasurement();
 									} else {
 										log.debug("connect failed");
